vkd3d: Fix sign vs unsigned compare warning.
UINT16 promotes to int rather than UINT here.

+++ libs/vkd3d/resource.c
@@ -4573,7 +4573,7 @@ static void vkd3d_create_texture_uav(struct d3d12_desc *descriptor,
                     key.u.texture.view_type = VK_IMAGE_VIEW_TYPE_3D;
                     key.u.texture.miplevel_idx = desc->Texture3D.MipSlice;
                     if (desc->Texture3D.FirstWSlice ||
-                        ((desc->Texture3D.WSize != max(1u, resource->desc.DepthOrArraySize >> desc->Texture3D.MipSlice)) &&
+                        ((desc->Texture3D.WSize != max(1u, (UINT)resource->desc.DepthOrArraySize >> desc->Texture3D.MipSlice)) &&
                             (desc->Texture3D.WSize != UINT_MAX)))
                     {
                         FIXME("Unhandled depth view %u-%u.\n",
